% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conventionals_LW.R
\name{Conventionals_LW}
\alias{Conventionals_LW}
\title{Analysis of conventionals data against LW standards and/or thresholds}
\usage{
Conventionals_LW(Chem_table, DU_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A dataframe containing analyzed conventionals data compared to LW
water quality criteria.
}
\description{
This function compares conventional water chemistry data against water quality
standards and/or thresholds for livestock watering (LW) use. For more
information, see the \href{https://www.env.nm.gov/surface-water-quality/calm/}{NMED Consolidated Assessment and Listing Methodology (CALM)} document.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed
df_Criteria <- NMwateR::example_criteria_processed

df_Conventionals_LW <- Conventionals_LW(Chem_table = df_Chem_combined
, DU_table = df_DU_processed
, Criteria_table = df_Criteria)
}
