% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{example_LANL_WQ_table}
\alias{example_LANL_WQ_table}
\title{Example LANL WQ Table}
\format{
A data frame with 15 rows and 900 columns:
\describe{
\item{AU ID}{Unique identifier for the Assessment Unit}
\item{AU / Waterbody Name}{Name of the waterbody or AU}
\item{Submitter / Data Source}{Organization or individual that submitted the data}
\item{Station ID}{Unique identifier for the monitoring station}
\item{Station Name}{Descriptive name of the monitoring station}
\item{Sample Date and Time}{Date and time when the sample was collected}
\item{Sample Media}{Type of sample collected}
\item{Sample Fraction}{Portion of the sample analyzed}
\item{Parameter Name}{Name of the analyte}
\item{Reported Value}{Numeric result of the analysis}
\item{Unit}{Measurement unit for the reported value}
\item{Sample Detection Limit (SDL)}{Minimum concentration that can be reliably detected by the analytical method}
\item{SDL Units}{Units associated with the detection limit}
\item{Less than detection limit?}{Indicator if the result is below the detection limit}
\item{Analytical Method}{Method or procedure used for analysis}
\item{Dilution Factor}{Factor by which the sample was diluted prior to analysis}
\item{Lab Name}{Name of the laboratory that performed the analysis}
\item{Analysis Date and Time}{Date and time when the analysis was completed}
\item{Lab Qualifier Code}{Laboratory-provided qualifier indicating data quality or conditions}
\item{Submitter Qualifier Code}{Submitter-provided qualifier for additional context or data quality notes}
\item{SSWQC Copper Acute (µg/L)}{State Surface Water Quality Criterion for acute copper toxicity}
\item{SSWQC Copper Chronic (µg/L)}{State Surface Water Quality Criterion for chronic copper toxicity}
}
}
\source{
Generated by \code{Load_Example_Files()} in \code{data-raw}.
}
\usage{
example_LANL_WQ_table
}
\description{
Example dataset containing LANL collected water quality data.
}
\keyword{datasets}
