% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{example_parameter_table}
\alias{example_parameter_table}
\title{Example Parameter Table}
\format{
A data frame with 133 rows and 17 columns:
\describe{
\item{CHR_UID}{Characteristic unique ID}
\item{CHR_NAME}{Characteristic name}
\item{WQS_NAME}{Water quality standard name}
\item{SortOrder}{Sort order for display}
\item{ShortName}{Short name for parameter}
\item{ShortNameOLD}{Previous short name}
\item{Units}{Units of measurement}
\item{USE}{Designated use category}
\item{ATTAINS_NAME}{ATTAINS name}
\item{ATTAINs_GROUP_ID}{ATTAINS group identifier}
\item{ATTAINS_GROUP_NAME}{ATTAINS group name}
\item{ATTAINS_ID}{ATTAINS identifier}
\item{ADB_NO}{ADB number}
\item{ADB_NAME}{ADB name}
\item{CAS}{CAS number}
\item{STORET}{STORET code}
\item{SHORT_CAUSE_NAME}{Short cause name}
}
}
\source{
Generated by \code{Load_Example_Files()} in \code{data-raw}.
}
\usage{
example_parameter_table
}
\description{
Example dataset containing parameters for SQUID analysis.
}
\keyword{datasets}
