% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{example_SQUID_LakeProfile_table}
\alias{example_SQUID_LakeProfile_table}
\title{Example SQUID Lake Profile Table}
\format{
A data frame with 644 rows and 14 columns:
\describe{
\item{PROJECT_NAME}{Project name}
\item{WATER_ID}{Waterbody identifier}
\item{MLOC_NAME}{Monitoring location name}
\item{MLOC_ID}{Monitoring location ID}
\item{ACT_ID}{Activity identifier}
\item{RES_DEPTH_HEIGHT}{Depth or height of measurement}
\item{CHARACTERISTIC}{Characteristic measured}
\item{MEASUREMENT}{Measurement value}
\item{UNITS}{Units of measurement}
\item{SE_START_DATE_TIME}{Sampling event start date/time}
\item{SWQB_QUALIFIER_CODE}{Qualifier code for SWQB}
\item{RES_COMMENTS}{Result comments}
\item{PRJ_UID}{Project unique ID}
\item{MLOC_UID}{Monitoring location unique ID}
}
}
\source{
Generated by \code{Load_Example_Files()} in \code{data-raw}.
}
\usage{
example_SQUID_LakeProfile_table
}
\description{
Example dataset for SQUID lake profile analysis.
}
\keyword{datasets}
