% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{example_SQUID_RStudio_table}
\alias{example_SQUID_RStudio_table}
\title{Example SQUID RStudio Table}
\format{
A data frame with 9265 rows and 46 columns:
\describe{
\item{WATER_ID}{Waterbody identifier}
\item{WATER_NAME}{Waterbody name}
\item{PROJECT_NAME}{Project name}
\item{PROJECT_VERIFIED}{Indicator if project verified}
\item{STATION}{Station code}
\item{STATION_NAME}{Station name}
\item{SAMPLING_EVENT_TYPE}{Type of sampling event}
\item{FIELD_STAFF}{Field staff name}
\item{ACT_ID}{Activity identifier}
\item{DATE}{Date of activity}
\item{TIME}{Time of activity}
\item{ACTIVITY_TYPE}{Type of activity}
\item{SAMPLE_TYPE}{Type of sample}
\item{ANALYTE_SUITE}{Analyte suite name}
\item{FLT_10UG}{Filter threshold indicator}
\item{SAMPLE_FRACTION}{Sample fraction type}
\item{CHR_UID}{Characteristic unique ID}
\item{CHARACTERISTIC_NAME}{Characteristic name}
\item{MEASUREMENT}{Measurement value}
\item{UNITS}{Units of measurement}
\item{SDL}{Sample detection limit}
\item{PQL}{Practical quantitation limit}
\item{MDL}{Method detection limit}
\item{MRL}{Minimum reporting level}
\item{SIGMA}{Standard deviation estimate}
\item{LESS_THAN_YN}{Flag if less than detection}
\item{GREATER_THAN_YN}{Flag if greater than threshold}
\item{ANLMTH_ID}{Analysis method ID}
\item{DILUTION_FACTOR}{Dilution factor applied}
\item{LAB_ID}{Laboratory identifier}
\item{ANALYSIS_DATE}{Date of analysis}
\item{WQX_QUALIFIER_CODE}{Qualifier code for WQX}
\item{LAB_QUALIFIER_CODE}{Laboratory qualifier code}
\item{SWQB_QUALIFIER_CODE}{SWQB qualifier code}
\item{RES_COMMENTS}{Result comments}
\item{ACTIVITY_COMMENTS}{Activity comments}
\item{COMMENTS}{General comments}
\item{CHR_CAS_NO}{CAS number for characteristic}
\item{ATTAINS_ID}{ATTAINS identifier}
\item{SE_TYPE_UID}{Sampling event type ID}
\item{RES_UID}{Result unique ID}
\item{ACT_UID}{Activity unique ID}
\item{ACTYP_UID}{Activity type unique ID}
\item{SE_UID}{Sampling event unique ID}
\item{MLOC_UID}{Monitoring location unique ID}
\item{PRJ_UID}{Project unique ID}
\item{AU_UID}{Assessment unit unique ID}
}
}
\source{
Generated by \code{Load_Example_Files()} in \code{data-raw}.
}
\usage{
example_SQUID_RStudio_table
}
\description{
Example dataset for SQUID RStudio integration.
}
\keyword{datasets}
