% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{Assessment of water quality data using NMED SWQS and CALM logic}
\usage{
assessment(
  Conventionals_ALU_table,
  Bacteria_PCR_SCR_table = NULL,
  Conventionals_LW_table = NULL,
  LTD_ALU_table = NULL,
  Nutrients_Lakes_table = NULL,
  Nutrients_Streams_table = NULL,
  pH_PCR_table = NULL,
  Salinity_IRR_table = NULL,
  SS_Copper_ALU_table = NULL,
  Toxics_ALU_nonHDM_table = NULL,
  Toxics_ALU_HDM_table = NULL,
  Toxics_DWS_table = NULL,
  Toxics_HH_table = NULL,
  Toxics_IRR_table = NULL,
  Toxics_LW_table = NULL,
  Toxics_WH_table = NULL,
  Turbidity_ALU_table = NULL
)
}
\arguments{
\item{Conventionals_ALU_table}{WQ analysis export from Conventionals_ALU() function.}

\item{Bacteria_PCR_SCR_table}{WQ analysis export from Bacteria_PCR_SCR() function.}

\item{Conventionals_LW_table}{WQ analysis export from Conventionals_LW() function.}

\item{LTD_ALU_table}{WQ analysis export from LTD_ALU() function.}

\item{Nutrients_Lakes_table}{WQ analysis export from Nutrients_Lakes() function.}

\item{Nutrients_Streams_table}{WQ analysis export from Nutrients_Streams() function.}

\item{pH_PCR_table}{WQ analysis export from pH_PCR() function.}

\item{Salinity_IRR_table}{WQ analysis export from Salinity_IRR() function.}

\item{SS_Copper_ALU_table}{WQ analysis export from SS_Copper_ALU() function.}

\item{Toxics_ALU_nonHDM_table}{WQ analysis export from Toxics_ALU_nonHDM() function.}

\item{Toxics_ALU_HDM_table}{WQ analysis export from Toxics_ALU_HDM() function.}

\item{Toxics_DWS_table}{WQ analysis export from Toxics_DWS() function.}

\item{Toxics_HH_table}{WQ analysis export from Toxics_HH() function.}

\item{Toxics_IRR_table}{WQ analysis export from Toxics_IRR() function.}

\item{Toxics_LW_table}{WQ analysis export from Toxics_LW() function.}

\item{Toxics_WH_table}{WQ analysis export from Toxics_WH() function.}

\item{Turbidity_ALU_table}{WQ analysis export from Turbidity_ALU() function.}
}
\value{
A list of four dataframes. Indiviual results are IR categories assigned
to each AU/DU/parameter combination. DU results are IR categories assigned to
each AU/DU combination. AU results are IR categories assigned to each AU. Lastly,
the upload report is for NMED to upload into SQUID to be transferred to ATTAINS.
}
\description{
This function uses the exports from the various water quality analyses and
assesses the results using logic from the New Mexico Surface Water Quality
Standards (SWQS) and the NMED Consolidated Assessment and Listing Methodology
(CALM) guidance manual.
}
\examples{
\dontrun{
assessment_list <- assessment(Conventionals_ALU_table = df_Conv_ALU
, Bacteria_PCR_SCR_table = df_Bacteria_PCR_SCR
, Conventionals_LW_table = df_Conventionals_LW
, LTD_ALU_table = df_LTD_ALU
, Nutrients_Lakes_table = df_Nutrients_Lakes
, Nutrients_Streams_table = df_Nutrients_Streams
, pH_PCR_table = df_pH_PCR
, Salinity_IRR_table = df_Salinity_IRR
, SS_Copper_ALU_table = df_SS_Copper_ALU
, Toxics_ALU_nonHDM_table = df_Tox_ALU_nHDM
, Toxics_ALU_HDM_table = df_Toxics_ALU_HDM
, Toxics_DWS_table = df_Toxics_DWS
, Toxics_HH_table = df_Toxics_HH
, Toxics_IRR_table = df_Toxics_IRR
, Toxics_LW_table = df_Toxics_LW
, Toxics_WH_table = df_Toxics_WH
, Turbidity_ALU_table = df_Turbidity_ALU)
df_Assess_Indiv_Res <- assessment_list$Assess_Indiv_Res
df_Assess_DU_Res <- assessment_list$Assess_DU_Res
df_Assess_AU_Res <- assessment_list$Assess_AU_Res
df_Assess_Upload_Report <- assessment_list$Assess_Upload_Report
}

}
