% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{Assessment of water quality data using NMED SWQS and CALM logic}
\usage{
assessment(
  Conventionals_ALU_table,
  Bacteria_PCR_SCR_table = NULL,
  Conventionals_LW_table = NULL,
  LTD_ALU_table = NULL,
  Nutrients_Lakes_table = NULL,
  Nutrients_Streams_table = NULL,
  pH_PCR_table = NULL,
  Salinity_IRR_table = NULL,
  SS_Copper_ALU_table = NULL,
  Toxics_ALU_nonHDM_table = NULL,
  Toxics_ALU_HDM_table = NULL,
  Toxics_DWS_table = NULL,
  Toxics_HH_table = NULL,
  Toxics_IRR_table = NULL,
  Toxics_LW_table = NULL,
  Toxics_WH_table = NULL,
  Turbidity_ALU_table = NULL
)
}
\arguments{
\item{Conventionals_ALU_table}{WQ analysis export from Conventionals_ALU() function.}

\item{Bacteria_PCR_SCR_table}{WQ analysis export from Bacteria_PCR_SCR() function.}

\item{Conventionals_LW_table}{WQ analysis export from Conventionals_LW() function.}

\item{LTD_ALU_table}{WQ analysis export from LTD_ALU() function.}

\item{Nutrients_Lakes_table}{WQ analysis export from Nutrients_Lakes() function.}

\item{Nutrients_Streams_table}{WQ analysis export from Nutrients_Streams() function.}

\item{pH_PCR_table}{WQ analysis export from pH_PCR() function.}

\item{Salinity_IRR_table}{WQ analysis export from Salinity_IRR() function.}

\item{SS_Copper_ALU_table}{WQ analysis export from SS_Copper_ALU() function.}

\item{Toxics_ALU_nonHDM_table}{WQ analysis export from Toxics_ALU_nonHDM() function.}

\item{Toxics_ALU_HDM_table}{WQ analysis export from Toxics_ALU_HDM() function.}

\item{Toxics_DWS_table}{WQ analysis export from Toxics_DWS() function.}

\item{Toxics_HH_table}{WQ analysis export from Toxics_HH() function.}

\item{Toxics_IRR_table}{WQ analysis export from Toxics_IRR() function.}

\item{Toxics_LW_table}{WQ analysis export from Toxics_LW() function.}

\item{Toxics_WH_table}{WQ analysis export from Toxics_WH() function.}

\item{Turbidity_ALU_table}{WQ analysis export from Turbidity_ALU() function.}
}
\value{
A list of four dataframes. Indiviual results are IR categories assigned
to each AU/DU/parameter combination. DU results are IR categories assigned to
each AU/DU combination. AU results are IR categories assigned to each AU. Lastly,
the upload report is for NMED to upload into SQUID to be transferred to ATTAINS.
}
\description{
This function uses the exports from the various water quality analyses and
assesses the results using logic from the New Mexico Surface Water Quality
Standards (SWQS) and the NMED Consolidated Assessment and Listing Methodology
(CALM) guidance manual.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_Criteria <- NMwateR::example_criteria_processed
df_DU_processed <- NMwateR::example_DU_processed
example_LANL_DU_table <- NMwateR::example_LANL_DU_table
example_LANL_WQ_table <- NMwateR::example_LANL_WQ_table

# Water quality analyses
## Conventionals ALU
df_Conv_ALU <- Conventionals_ALU(Chem_table = df_Chem_combined
                                 , DU_table = df_DU_processed)

## Toxics ALU (nonHDM)
# Hardness-dependent metals excluded
df_Tox_ALU_nHDM <- Toxics_ALU_nonHDM(Chem_table = df_Chem_combined
                                     , DU_table = df_DU_processed
                                     , Criteria_table = df_Criteria)

## Toxics ALU (HDM)
# Hardness-dependent metals only
Toxics_ALU_HDM_list <- Toxics_ALU_HDM(Chem_table = df_Chem_combined
                                      , DU_table = df_DU_processed
                                      , Criteria_table = df_Criteria)


df_Toxics_ALU_HDM <- Toxics_ALU_HDM_list$Toxics_ALU_HDM

# cleanup
rm(Toxics_ALU_HDM_list)

## Conventionals LW
df_Conventionals_LW <- Conventionals_LW(Chem_table = df_Chem_combined
                                        , DU_table = df_DU_processed
                                        , Criteria_table = df_Criteria)

## Salinity IRR ####
df_Salinity_IRR <- Salinity_IRR(Chem_table = df_Chem_combined
                                , DU_table = df_DU_processed)

## Toxics HH ####
df_Toxics_HH <- Toxics_HH(Chem_table = df_Chem_combined
                          , Criteria_table = df_Criteria)

## Site-specific copper
#Only used for LANL data
SS_Copper_ALU_list <- SS_Copper_ALU(DU_LANL_Stations_table = example_LANL_DU_table
                                    , LANL_WQ_data = example_LANL_WQ_table)

df_SS_Copper_ALU <- SS_Copper_ALU_list$df_SS_Copper_ALU

rm(SS_Copper_ALU_list, example_LANL_DU_table, example_LANL_WQ_table)

## Toxics DWS
df_Toxics_DWS <- Toxics_DWS(Chem_table = df_Chem_combined
                            , Criteria_table = df_Criteria)

## Toxics IRR
df_Toxics_IRR <- Toxics_IRR(Chem_table = df_Chem_combined
                            , Criteria_table = df_Criteria)

## Toxics WH
df_Toxics_WH <- Toxics_WH(Chem_table = df_Chem_combined
                         , Criteria_table = df_Criteria)

## Toxics LW
df_Toxics_LW <- Toxics_LW(Chem_table = df_Chem_combined
                          , Criteria_table = df_Criteria)

## Bacteria PCR/SCR
df_Bacteria_PCR_SCR <- Bacteria_PCR_SCR(Chem_table = df_Chem_combined
                                        , DU_table = df_DU_processed)

## pH PCR
df_pH_PCR <- pH_PCR(Chem_table = df_Chem_combined
                    , DU_table = df_DU_processed)

## Turbidity ALU
Turbidity_ALU_list <- Turbidity_ALU(Chem_table = df_Chem_combined
                                    , DU_table = df_DU_processed
                                   , Criteria_table = df_Criteria)

df_Turbidity_ALU <- Turbidity_ALU_list$Turbidity_ALU

## LTD ALU
df_LTD_ALU <- LTD_ALU(Chem_table = df_Chem_combined
                      , DU_table = df_DU_processed)

## Nutrients (Lakes)
Nutrients_Lakes_list <- Nutrients_Lakes(Chem_table = df_Chem_combined
                                        , DU_table = df_DU_processed)

df_Nutrients_Lakes <- Nutrients_Lakes_list$Nutrients_Lakes

# cleanup
rm(Nutrients_Lakes_list)

## Nutrients (Streams) ####
Nutrients_Streams_list <- Nutrients_Streams(Chem_table = df_Chem_combined
                                            , DU_table = df_DU_processed)

df_Nutrients_Streams <- Nutrients_Streams_list$Nutrients_Streams

# cleanup
rm(Nutrients_Streams_list)

# Assessment ####
assessment_list <- assessment(Conventionals_ALU_table = df_Conv_ALU
                              , Bacteria_PCR_SCR_table = df_Bacteria_PCR_SCR
                              , Conventionals_LW_table = df_Conventionals_LW
                              , LTD_ALU_table = df_LTD_ALU
                              , Nutrients_Lakes_table = df_Nutrients_Lakes
                              , Nutrients_Streams_table = df_Nutrients_Streams
                              , pH_PCR_table = df_pH_PCR
                              , Salinity_IRR_table = df_Salinity_IRR
                              , SS_Copper_ALU_table = df_SS_Copper_ALU
                              , Toxics_ALU_nonHDM_table = df_Tox_ALU_nHDM
                              , Toxics_ALU_HDM_table = df_Toxics_ALU_HDM
                              , Toxics_DWS_table = df_Toxics_DWS
                              , Toxics_HH_table = df_Toxics_HH
                              , Toxics_IRR_table = df_Toxics_IRR
                              , Toxics_LW_table = df_Toxics_LW
                              , Toxics_WH_table = df_Toxics_WH
                              , Turbidity_ALU_table = df_Turbidity_ALU)
df_Assess_Indiv_Res <- assessment_list$Assess_Indiv_Res
df_Assess_DU_Res <- assessment_list$Assess_DU_Res
df_Assess_AU_Res <- assessment_list$Assess_AU_Res
df_Assess_Upload_Report <- assessment_list$Assess_Upload_Report

}
