% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bacteria_PCR_SCR.R
\name{Bacteria_PCR_SCR}
\alias{Bacteria_PCR_SCR}
\title{Analysis of bacteria data against PCR and SCR standards}
\usage{
Bacteria_PCR_SCR(Chem_table, DU_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}
}
\value{
A dataframe containing analyzed bacteria data compared to PCR and SCR
water quality criteria.
}
\description{
This function compares bacteria data against water quality
standards for primary contact recreation (PCR) and secondary contact recreation
(SCR) uses. For more information, see the NMED Consolidated Assessment and
Listing Methodology (CALM) guidance manual.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed

df_Bacteria_PCR_SCR <- Bacteria_PCR_SCR(Chem_table = df_Chem_combined
, DU_table = df_DU_processed)

}
