% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conventionals_ALU.R
\name{Conventionals_ALU}
\alias{Conventionals_ALU}
\title{Analysis of conventionals data against ALU standards}
\usage{
Conventionals_ALU(Chem_table, DU_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}
}
\value{
A dataframe containing analyzed conventionals data compared to ALU
water quality criteria.
}
\description{
This function compares conventional water chemistry data against water quality
standards for aquatic life use (ALU). For more information, see the \href{https://www.env.nm.gov/surface-water-quality/calm/}{NMED Consolidated Assessment and Listing Methodology (CALM)} document.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed
df_Conv_ALU <- Conventionals_ALU(Chem_table = df_Chem_combined
, DU_table = df_DU_processed)

}
