% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conventionals_LW.R
\name{Conventionals_LW}
\alias{Conventionals_LW}
\title{Analysis of conventionals data against LW standards}
\usage{
Conventionals_LW(Chem_table, DU_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A dataframe containing analyzed conventionals data compared to LW
water quality criteria.
}
\description{
This function compares conventional water chemistry data against water quality
standards for livestock watering (LW) use. For more information, see the NMED
Consolidated Assessment and Listing Methodology (CALM) guidance manual.
}
\examples{
\dontrun{
df_Conventionals_LW <- Conventionals_LW(Chem_table = df_Chem_combined
, DU_table = df_DU_processed
, Criteria_table = df_Criteria)}
}
