% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Prep.R
\name{Data_Prep}
\alias{Data_Prep}
\title{Compile data from SQUID for water quality analyses}
\usage{
Data_Prep(
  criteria_table,
  parameter_table,
  SQUID_RStudio_table,
  SQUID_DU_table,
  SQUID_LTD_table = NULL,
  SQUID_LakeProfile_table = NULL
)
}
\arguments{
\item{criteria_table}{Criteria table is output from SQUID. Filters water quality
data to those in the criteria table. Has criteria to compare data against.}

\item{parameter_table}{Parameter table is output from SQUID. Filters water quality
data to those in the parameter table.}

\item{SQUID_RStudio_table}{RStudio query from SQUID. Entirely grab data from SQUID
projects.}

\item{SQUID_DU_table}{Stations DU table from SQUID. Has DU and AU specific water
quality criteria which are used in some cases.}

\item{SQUID_LTD_table}{Optional long-term deployment summary table from SQUID.
Not raw data.}

\item{SQUID_LakeProfile_table}{Optional lake profile table from SQUID.}
}
\value{
A dataframe containing compiled and quality controlled water quality
necessary for subsequent analyses.
}
\description{
This function compiles multiple tables from SQUID into a single output used for
subsequent water quality analysis functions. The LTD and lake profile datasets
are optional, whereas the RStudio, stations DU, parameter, and criteria tables
are required for the function to run.
}
\examples{
example_criteria_table <- NMwateR::example_criteria_table
example_parameter_table <- NMwateR::example_parameter_table
example_SQUID_RStudio_table <- NMwateR::example_SQUID_RStudio_table
example_SQUID_DU_table <- NMwateR::example_SQUID_DU_table
example_SQUID_LTD_table <- NMwateR::example_SQUID_LTD_table
example_SQUID_LakeProfile_table <- NMwateR::example_SQUID_LakeProfile_table

my_data_list <- Data_Prep(criteria_table = example_criteria_table
                          , parameter_table = example_parameter_table
                          , SQUID_RStudio_table = example_SQUID_RStudio_table
                          , SQUID_DU_table = example_SQUID_DU_table
                          , SQUID_LTD_table = example_SQUID_LTD_table
                          , SQUID_LakeProfile_table = example_SQUID_LakeProfile_table)
# cleanup
rm(example_criteria_table, example_parameter_table, example_SQUID_RStudio_table
 , example_SQUID_DU_table, example_SQUID_LTD_table, example_SQUID_LakeProfile_table)

df_Chem_combined <- my_data_list$Chem_Combined
df_DU_processed <- my_data_list$DU_Processed
df_Criteria <- my_data_list$Criteria_Formatted

}
