% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTD_ALU.R
\name{LTD_ALU}
\alias{LTD_ALU}
\title{Analysis of long-term deployment data against ALU standards}
\usage{
LTD_ALU(Chem_table, DU_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}
}
\value{
A dataframe containing analyzed conventional LTD data compared to ALU
water quality criteria.
}
\description{
This function compares conventional long-term deployment water chemistry data
against water quality standards for aquatic life use (ALU). For more
information, see the \href{https://www.env.nm.gov/surface-water-quality/calm/}{NMED Consolidated Assessment and Listing Methodology (CALM)} document.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed

df_LTD_ALU <- LTD_ALU(Chem_table = df_Chem_combined
, DU_table = df_DU_processed)

}
