% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_PCR.R
\name{pH_PCR}
\alias{pH_PCR}
\title{Analysis of pH data against PCR standards}
\usage{
pH_PCR(Chem_table, DU_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}
}
\value{
A dataframe containing analyzed pH data compared to PCR
water quality criteria.
}
\description{
This function compares pH data against water quality standards for
primary contact recreation (PCR) use. For more information, see the NMED
Consolidated Assessment and Listing Methodology (CALM) guidance manual.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed

df_pH_PCR <- pH_PCR(Chem_table = df_Chem_combined
, DU_table = df_DU_processed)

}
