% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salinity_IRR.R
\name{Salinity_IRR}
\alias{Salinity_IRR}
\title{Analysis of salinity data against IRR standards and/or thresholds}
\usage{
Salinity_IRR(Chem_table, DU_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}
}
\value{
A dataframe containing analyzed salinity data compared to IRR
water quality criteria.
}
\description{
This function compares salinity data against water quality
standards and/or thresholds for irrigation (IRR) use. For more information, see the \href{https://www.env.nm.gov/surface-water-quality/calm/}{NMED Consolidated Assessment and Listing Methodology (CALM)} document.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed

df_Salinity_IRR <- Salinity_IRR(Chem_table = df_Chem_combined
, DU_table = df_DU_processed)

}
