% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toxics_ALU_nonHDM.R
\name{Toxics_ALU_nonHDM}
\alias{Toxics_ALU_nonHDM}
\title{Analysis of non hardness dependent toxics data against ALU standards}
\usage{
Toxics_ALU_nonHDM(Chem_table, DU_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A dataframe containing analyzed non hardness dependent toxics data
compared to ALU water quality criteria.
}
\description{
This function compares non hardness dependent toxics data against water quality
standards for aquatic life use (ALU). For more information, see the NMED
Consolidated Assessment and Listing Methodology (CALM) guidance manual.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_DU_processed <- NMwateR::example_DU_processed
df_Criteria <- NMwateR::example_criteria_processed

df_Tox_ALU_nHDM <- Toxics_ALU_nonHDM(Chem_table = df_Chem_combined
, DU_table = df_DU_processed
, Criteria_table = df_Criteria)
}
