% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toxics_HH.R
\name{Toxics_HH}
\alias{Toxics_HH}
\title{Analysis of toxics data against HH standards for ALU}
\usage{
Toxics_HH(Chem_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A dataframe containing analyzed toxics data compared to HH
water quality criteria.
}
\description{
This function compares toxics data against water quality standards for
human health (HH) under the aquatic life use (ALU). For more information,
see the NMED Consolidated Assessment and Listing Methodology (CALM)
guidance manual.
}
\examples{
\dontrun{
df_Toxics_HH <- Toxics_HH(Chem_table = df_Chem_combined
, Criteria_table = df_Criteria)}

}
