% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toxics_LW.R
\name{Toxics_LW}
\alias{Toxics_LW}
\title{Analysis of toxics data against LW standards}
\usage{
Toxics_LW(Chem_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A dataframe containing analyzed toxics data compared to LW
water quality criteria.
}
\description{
This function compares toxics data against water quality
standards for livestock watering (LW) use. For more information, see the NMED
Consolidated Assessment and Listing Methodology (CALM) guidance manual.
}
\examples{
\dontrun{
df_Toxics_LW <- Toxics_LW(Chem_table = df_Chem_combined
, Criteria_table = df_Criteria)}
}
