% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turbidity_ALU.R
\name{Turbidity_ALU}
\alias{Turbidity_ALU}
\title{Analysis of turbidity data against ALU standards and/or thresholds}
\usage{
Turbidity_ALU(Chem_table, DU_table, Criteria_table)
}
\arguments{
\item{Chem_table}{Compiled water chemistry data from Data_Prep function. Usually
contains a combination of grab, LTD, and lake profile data.}

\item{DU_table}{Quality controlled Stations DU table from Data_Prep function.}

\item{Criteria_table}{Quality controlled criteria table from Data_Prep function.}
}
\value{
A list of two dataframes. The first contains analyzed turbidity data
compared to ALU water quality criteria. The second, labeled "Indiv_Res"
is an intermediate file used for QA/QC purposes.
}
\description{
This function compares turbidity data against water quality
standards and/or thresholds for aquatic life use (ALU). For more information, see the
\href{https://www.env.nm.gov/surface-water-quality/calm/}{NMED Consolidated Assessment and Listing Methodology (CALM)} document.
}
\examples{
df_Chem_combined <- NMwateR::example_chemistry_processed
df_Criteria <- NMwateR::example_criteria_processed
df_DU_processed <- NMwateR::example_DU_processed

Turbidity_ALU_list <- Turbidity_ALU(Chem_table = df_Chem_combined
, DU_table = df_DU_processed
, Criteria_table = df_Criteria)

df_Turbidity_ALU <- Turbidity_ALU_list$Turbidity_ALU
df_Turbidity_ALU_Indiv_Res <- Turbidity_ALU_list$Turbidity_ALU_Indiv_Res
}
